Remove unused code and update tests for GestionPermission
