+++ app/app/Models/Autorisation/Permission.php
@@ -15,5 +15,4 @@ class Permission extends Model
         'created_at',
         'updated_at'
     ];
-    
 }
